+++ Students.h
@@ -45,9 +45,7 @@ void InsertMark(string student_name, string subject_name, Mark mark) {
 
 void RemoveMark(string student_name, string subject_name, Mark mark) {
     auto& marks = _students[student_name][subject_name];
-    auto mark_it = remove(marks.begin(), marks.end(), mark);
-    marks.erase(mark_it, marks.end());
-//    marks.erase(remove(marks.begin(), marks.end(), mark), marks.end());
+    marks.erase(remove(marks.begin(), marks.end(), mark), marks.end());
 }
 
 void InsertMarks(string student_name, string subject_name, Marks marks) {
@@ -61,47 +59,3 @@ void UpdateMark(string student_name, string subject_name, Mark old_mark, Mark ne
     replace(marks.begin(), marks.end(), old_mark, new_mark);
 }
 
-void ExportData(string filename) {
-    ofstream outfile(filename);
-    for (auto& student : _students) {
-        outfile << student.first << endl;
-        for (auto& subject : student.second) {
-            outfile << subject.first << endl;
-            for (auto& mark : subject.second) {
-                outfile << mark << ",";
-            }
-            outfile << endl;
-        }
-        outfile << endl;
-    }
-    outfile.close();
-    cout << "Данные экспортированы в файл " << filename << endl;
-}
-
-void ImportData(string filename) {
-    ifstream infile(filename);
-    string line;
-    string student_name;
-    string subject_name;
-    Mark mark;
-
-    while (getline(infile, line)) {
-        if (line.empty()) {
-            continue;
-        }
-        if (line.back() == ':') {
-            student_name = line.substr(0, line.size() - 1);
-            InsertStudent(student_name);
-        } else if (isalpha(line[0])) {
-            subject_name = line;
-            InsertSubject(student_name, subject_name);
-        } else {
-            stringstream ss(line);
-            while (getline(ss, mark, ',')) {
-                InsertMark(student_name, subject_name, mark);
-            }
-        }
-    }
-    infile.close();
-    cout << "Данные импортированы из файла " << filename << endl;
-}
